# docs/swagger.yaml
definitions:
...
  database.AccessToken:
    properties:
      application:
        allOf:
        - $ref: '#/definitions/types.AccessTokenApp'
        description: 'example: csghub, starship'
      created_at:
        type: string
      expired_at:
        type: string
      git_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      permission:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
    type: object
...
  database.Collection:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      likes:
        type: integer
      name:
        type: string
      namespace:
        type: string
      nickname:
        type: string
      private:
        type: boolean
      repositories:
        items:
          $ref: '#/definitions/database.Repository'
        type: array
      theme:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  database.Mirror:
    properties:
      created_at:
        type: string
      id:
        type: integer
      interval:
        type: string
      last_message:
        type: string
      last_updated_at:
        type: string
      local_repo_path:
        type: string
      mirror_source:
        $ref: '#/definitions/database.MirrorSource'
      mirror_source_id:
        type: integer
      mirror_task_id:
        type: integer
      next_execution_timestamp:
        type: string
      priority:
        $ref: '#/definitions/types.MirrorPriority'
      progress:
        type: integer
      push_mirror_created:
        type: boolean
      repository:
        $ref: '#/definitions/database.Repository'
      repository_id:
        type: integer
      source_repo_path:
        type: string
      source_url:
        type: string
      status:
        $ref: '#/definitions/types.MirrorTaskStatus'
      updated_at:
        type: string
    type: object
  database.MirrorSource:
    properties:
      created_at:
        type: string
      id:
        type: integer
      info_api_url:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
...
  database.Namespace:
    properties:
      created_at:
        type: string
      id:
        type: integer
      mirrored:
        type: boolean
      namespace_type:
        $ref: '#/definitions/database.NamespaceType'
      path:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
    type: object
  database.NamespaceType:
    enum:
    - user
    - organization
    type: string
    x-enum-varnames:
    - UserNamespace
    - OrgNamespace
...
  database.Repository:
    properties:
      created_at:
        type: string
      default_branch:
        type: string
      description:
        type: string
      download_count:
        type: integer
      downloads:
        items:
          $ref: '#/definitions/database.RepositoryDownload'
        type: array
      git_path:
        type: string
      http_clone_url:
        type: string
      id:
        type: integer
      labels:
        description: Depreated
        type: string
      license:
        type: string
      likes:
        type: integer
      mirror:
        $ref: '#/definitions/database.Mirror'
      name:
        type: string
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      readme:
        description: Depreated
        type: string
      repository_type:
        $ref: '#/definitions/types.RepositoryType'
      sensitive_check_status:
        $ref: '#/definitions/types.SensitiveCheckStatus'
      source:
        $ref: '#/definitions/types.RepositorySource'
      ssh_clone_url:
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/database.Tag'
        type: array
      updated_at:
        type: string
      user:
        $ref: '#/definitions/database.User'
      user_id:
        type: integer
    type: object
  database.RepositoryDownload:
    properties:
      click_download_count:
        type: integer
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      repository:
        $ref: '#/definitions/database.Repository'
      repository_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  database.Tag:
    properties:
      built_in:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      group:
        type: string
      id:
        type: integer
      name:
        type: string
      scope:
        $ref: '#/definitions/database.TagScope'
      show_name:
        type: string
      updated_at:
        type: string
    type: object
  database.TagScope:
    enum:
    - model
    - dataset
    - code
    - space
    - prompt
    type: string
    x-enum-varnames:
    - ModelTagScope
    - DatasetTagScope
    - CodeTagScope
    - SpaceTagScope
    - PromptTagScope
  database.User:
    properties:
      accessTokens:
        items:
          $ref: '#/definitions/database.AccessToken'
        type: array
      avatar:
        type: string
      bio:
        type: string
      can_change_username:
        description: allow user to change username once
        type: boolean
      company_verified:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      gender:
        type: string
      git_id:
        type: integer
      homepage:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      name:
        type: string
      namespace:
        items:
          $ref: '#/definitions/database.Namespace'
        type: array
      password_hash:
        description: password for user registered without casdoor
        type: string
      phone:
        type: string
      phone_verified:
        type: boolean
      reg_provider:
        description: |-
          user registered from default login page, from casdoor, etc. Possible values:

          - "default"
          - "casdoor"
        type: string
      role_mask:
        type: string
      updated_at:
        type: string
      username:
        type: string
      uuid:
        description: TODO:add unique index after migration
        type: string
    type: object
...
  types.APIBadRequest:
    type: object
  types.APIInternalServerError:
    type: object
...
  types.AccessTokenApp:
    enum:
    - git
    - git
    - mirror
    - starship
    type: string
    x-enum-varnames:
    - AccessTokenAppGit
    - AccessTokenAppCSGHub
    - AccessTokenAppMirror
    - AccessTokenAppStarship
...
  types.Collection:
    properties:
      avatar:
        type: string
      can_manage:
        type: boolean
      can_write:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      likes:
        type: integer
      name:
        type: string
      namespace:
        type: string
      nickname:
        type: string
      private:
        type: boolean
      repositories:
        items:
          $ref: '#/definitions/types.CollectionRepository'
        type: array
      theme:
        type: string
      updated_at:
        type: string
      user_likes:
        type: boolean
      username:
        type: string
    type: object
  types.CollectionRepository:
    properties:
      created_at:
        type: string
      description:
        type: string
      download_count:
        type: integer
      id:
        type: integer
      likes:
        type: integer
      name:
        type: string
      nickname:
        type: string
      path:
        type: string
      private:
        type: boolean
      repository_type:
        $ref: '#/definitions/types.RepositoryType'
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  types.CreateCollectionReq:
    properties:
      description:
        type: string
      name:
        example: collection1
        type: string
      namespace:
        example: user_or_org_name
        type: string
      nickname:
        example: collection nick name
        type: string
      private:
        type: boolean
      theme:
        example: '#fff000'
        type: string
    type: object
...
  types.MirrorPriority:
    enum:
    - 3
    - 2
    - 1
    type: integer
    x-enum-varnames:
    - HighMirrorPriority
    - MediumMirrorPriority
    - LowMirrorPriority
...
  types.MirrorTaskStatus:
    enum:
    - waiting
    - running
    - repo_synced
    - finished
    - failed
    - incomplete
    type: string
    x-enum-varnames:
    - MirrorWaiting
    - MirrorRunning
    - MirrorRepoSynced
    - MirrorFinished
    - MirrorFailed
    - MirrorIncomplete
...
  types.RepoTag:
    properties:
      built_in:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      group:
        type: string
      name:
        type: string
      show_name:
        type: string
      updated_at:
        type: string
    type: object
...
  types.RepositorySource:
    enum:
    - opencsg
    - local
    - huggingface
    type: string
    x-enum-varnames:
    - OpenCSGSource
    - LocalSource
    - HuggingfaceSource
  types.RepositorySyncStatus:
    enum:
    - pending
    - inprogress
    - failed
    - completed
    type: string
    x-enum-varnames:
    - SyncStatusPending
    - SyncStatusInProgress
    - SyncStatusFailed
    - SyncStatusCompleted
  types.RepositoryType:
    enum:
    - model
    - dataset
    - space
    - code
    - prompt
    - ""
    type: string
    x-enum-varnames:
    - ModelRepo
    - DatasetRepo
    - SpaceRepo
    - CodeRepo
    - PromptRepo
    - UnknownRepo
...
  types.Response:
    properties:
      data: {}
      msg:
        type: string
    type: object
  types.ResponseWithTotal:
    properties:
      data: {}
      msg:
        type: string
      total:
        type: integer
    type: object
...
  types.SensitiveCheckStatus:
    enum:
    - -1
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      SensitiveCheckException: error happen
      SensitiveCheckFail: sensitive content detected
      SensitiveCheckPass: pass
      SensitiveCheckPending: default
      SensitiveCheckSkip: skip
    x-enum-varnames:
    - SensitiveCheckFail
    - SensitiveCheckPending
    - SensitiveCheckPass
    - SensitiveCheckSkip
    - SensitiveCheckException
...
  types.UpdateCollectionReposReq:
    properties:
      repo_ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /collections:
    get:
      consumes:
      - application/json
      description: get all collections
      parameters:
      - description: search text
        in: query
        name: search
        type: string
      - default: '"trending"'
        description: sort by
        in: query
        name: sort
        type: string
      - default: 50
        description: per
        in: query
        name: per
        type: integer
      - default: 1
        description: per page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Collection'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      summary: get all collections
      tags:
      - Collection
    post:
      consumes:
      - application/json
      description: create a collection
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateCollectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Collection'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: create a collection
      tags:
      - Collection
  /collections/{id}:
    delete:
      consumes:
      - application/json
      description: delete a exists collection
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: Delete a exists collection
      tags:
      - Collection
    get:
      consumes:
      - application/json
      description: get a collection detail
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Collection'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      summary: get a collection detail
      tags:
      - Collection
    put:
      consumes:
      - application/json
      description: update a collection
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateCollectionReq'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Collection'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: update a collection
      tags:
      - Collection
  /collections/{id}/repos:
    delete:
      consumes:
      - application/json
      description: remove repos from a collection
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCollectionReposReq'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Collection'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: remove repos from a collection
      tags:
      - Collection
    post:
      consumes:
      - application/json
      description: Add repos to a collection
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateCollectionReposReq'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/database.Collection'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: Add repos to a collection
      tags:
      - Collection
...
securityDefinitions:
  ApiKey:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"